[editor]: improve how insertion of parens is handled (#36)
In particular, instead of inserting a specific parens node we now insert ( and ) glyphs. The matching paren to the one inserted by the user will appear at the opposite end of the row where the cursor is and will be in a pending state until its exact location is confirmed by the user.