[editor]: refactor Editor.Node to be more flexible (#100)
In particular Editor.Node takes two params: one for the type to store in atoms and the other to control which properties are mixed into each of the node types. This is more flexible than setting type of the id prop and will allow us to have ids, locations, or in some cases no extra properties. I'll add location data in a future PR.