Refactor color styles (#387)
Layout nodes now have a `style` property which is an object can contain `color` and `cancel` properties.  SceneGraph nodes have a similar `style` property which can contain `fill` and `stroke`.  There is also a top-level `<g>` surrounding the content layer in the Renderer which sets `fill` and `stroke` to `"currentColor"` so that we don't have to do that on every node that doesn't explicitly have a color.